test: Combine test suites
